+++ src/test/resources/model/modelParentWithoutTableProperty.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App {
+class DemoOneButActuallyUsers extends \Illuminate\Database\Eloquent\Model {
+}
+class User extends DemoOneButActuallyUsers {
+}
+}
+
+(new \App\User())->newQuery()->get('<caret>');